requirements.txt
fbchat-muqit
pillow
requests
bs4
dataset
rich
flask
yt-dlp[default]
